move SamplingResult to base_sampler.py
